chore(tabs): make the scroll-settle helper local to the spec
Its only caller is the tabs e2e spec, so it moves there instead of living
in the shared utils barrel. The selector parameter goes with it: it was
always the tab scrollport, which is now declared next to the helper.
