+++ packages/e2e/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './autofill';
 export * from './focus-scroll';
 export * from './overflow-shadow';
-export * from './scroll-position';
 export * from './scrollbar';
 export * from './theme';
